move Location to separate file
